feat: refactroing a bit the photoVideo component
