+++ src/components/PhotoVideo/index.tsx
@@ -1,21 +1,16 @@
 import { useRef, useState } from 'react';
 import { Box, Link, Container, Grid, Stack, Typography } from '@mui/material';
 import PlayArrow from '@mui/icons-material/PlayArrow';
-// import { type Generic } from 'types/components';
-import { isJSX } from '../utils';
-import YouTube, { type YouTubePlayer } from 'react-youtube';
-import { type CommonProps } from 'types/components';
+import { isJSX } from '../../utils';
+import YouTubeVideo, { type PhotoVideoYouTubeProps } from './YouTube';
+import Text from './Text';
 
-export interface PhotoVideoProps extends YProps {
+export interface PhotoVideoProps extends PhotoVideoYouTubeProps {
   src: string;
   altText?: string;
-  title?: string;
-  subtitle?: string;
-  playButtonLabel?: string;
   caption?: string;
   captionAlign?: 'left' | 'center';
   maxHeight?: string;
-  full?: boolean;
 }
 
 const youtubeParser = (url: string) => {
@@ -23,124 +18,6 @@ const youtubeParser = (url: string) => {
     /^.*((youtu.be\/)|(v\/)|(\/u\/\w\/)|(embed\/)|(watch\?))\??v?=?([^#&?]*).*/;
   const match = url.match(regExp);
   return match && match[7].length === 11 ? match[7] : false;
-};
-
-interface YProps extends Ytext {
-  full?: boolean;
-  reverse?: boolean;
-  youtubeID: string;
-  autoplay?: boolean;
-  playButtonLabel?: string;
-}
-
-interface Ytext extends CommonProps {
-  title?: string;
-  subtitle?: string;
-}
-
-const Text = ({ title, subtitle, theme = 'dark' }: Ytext) => {
-  const color = theme === 'dark' ? 'primary.contrastText' : 'text.primary';
-  return (
-    <>
-      {title && (
-        <Typography variant="h5" mb={4} color={color}>
-          {title}
-        </Typography>
-      )}
-      {subtitle && (
-        <Typography paragraph sx={{ fontSize: '16px' }} mb={3} color={color}>
-          {subtitle}
-        </Typography>
-      )}
-    </>
-  );
-};
-
-const YouTubeVideo = ({
-  title,
-  subtitle,
-  full,
-  reverse,
-  theme,
-  youtubeID,
-  autoplay = false,
-  playButtonLabel = 'Guarda il video',
-}: YProps) => {
-  const videoRef = useRef<YouTubePlayer>(null);
-
-  const play = async (e: React.MouseEvent) => {
-    try {
-      e.preventDefault();
-      videoRef.current.playVideo();
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  return (
-    <Box
-      component="section"
-      bgcolor={theme === 'dark' ? 'primary.dark' : 'white'}
-      py={7}
-    >
-      <Container maxWidth="lg">
-        <Grid
-          container
-          spacing={{ xs: 3, md: 16 }}
-          alignItems={'center'}
-          direction={reverse ? 'row-reverse' : 'row'}
-        >
-          <Grid item xs={12} md={full ? 12 : 6}>
-            <YouTube
-              videoId={youtubeID}
-              id={`video-${youtubeID}`}
-              className="video__frame"
-              onReady={({ target }) => {
-                videoRef.current = target;
-              }}
-              style={{
-                aspectRatio: 1.777,
-                borderRadius: '25px',
-                overflow: 'hidden',
-              }}
-              opts={{
-                width: '100%',
-                height: '100%',
-                host: 'https://www.youtube-nocookie.com',
-                playerVars: {
-                  autoplay: autoplay ? 1 : 0,
-                  showinfo: 0,
-                  mute: 1,
-                },
-              }}
-            />
-          </Grid>
-          {!full && (
-            <Grid item md={6}>
-              <Text title={title} subtitle={subtitle} theme={theme} />
-              <Stack direction="row" alignItems="center">
-                <Link
-                  component="button"
-                  onClick={play}
-                  sx={{
-                    fontWeight: 700,
-                    fontSize: '16px',
-                    textDecoration: 'none',
-                    color:
-                      theme === 'dark'
-                        ? 'primary.contrastText'
-                        : 'primary.dark',
-                  }}
-                >
-                  {playButtonLabel}
-                </Link>
-              </Stack>
-            </Grid>
-          )}
-        </Grid>
-      </Container>
-    </Box>
-  );
 };
 
 const PhotoVideo = (props: PhotoVideoProps) => {
@@ -159,7 +36,7 @@ const PhotoVideo = (props: PhotoVideoProps) => {
     theme = 'dark',
   } = props;
   const [playing, setPlaying] = useState(autoplay);
-  const videoRef = useRef<HTMLVideoElement | YouTubePlayer>(null);
+  const videoRef = useRef<HTMLVideoElement>(null);
 
   const scrIsVideo =
     !isJSX(src) && (/^.*\.(mp4|webm|ogg)$/.test(src) || youtubeParser(src));
@@ -171,10 +48,8 @@ const PhotoVideo = (props: PhotoVideoProps) => {
       e.preventDefault();
       setPlaying(!playing);
       if (playing) {
-        if (isYoutubeSrc) return videoRef.current.pauseVideo();
-        await videoRef.current?.pause();
+        videoRef.current?.pause();
       } else {
-        if (isYoutubeSrc) return videoRef.current.playVideo();
         await videoRef.current?.play();
       }
     } catch (error) {
